+++ src/lib/core/channels/noise-channel.js
@@ -1,9 +1,6 @@
 import { LENGTH_COUNTER_VALUES } from "../common/constants";
 import { logger }                from "../utils/logger";
 
-const TIMER_PERIODS_NTSC = [ 4, 8, 16, 32, 64, 96, 128, 160, 202, 254, 380, 508, 762, 1016, 2034, 4068 ];
-const TIMER_PERIODS_PAL  = [ 4, 8, 14, 30, 60, 88, 118, 148, 188, 236, 354, 472, 708,  944, 1890, 3778 ];
-
 //=========================================================
 // Noise channel
 //=========================================================
@@ -29,8 +26,8 @@ export class NoiseChannel {
         }
     }
 
-    setNTSCMode(ntscMode) {
-        this.timerPeriods = ntscMode ? TIMER_PERIODS_NTSC : TIMER_PERIODS_PAL;
+    setRegionParams(params) {
+        this.timerPeriods = params.noiseChannelTimerPeriods;
     }
 
     //=========================================================