+++ CHANGELOG.md
@@ -3,6 +3,8 @@ All notable changes to this project will be documented in this file.
 This project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased][unreleased]
+### Changed
+- *TV system* configuration option renamed to *Region*.
 
 ## [0.2.0] - 2015-05-18
 ### Fixed